pyproject.toml
[project]
name = "llm"
version = "0.1.0"
requires-python = ">3.7"
readme = "README.md"
authors = [{ name = "Team", email = "team@team.com" }]

dependencies = [
  # Data processing libraries
  "pandas",
  "polars",
  # Api server libraries
  "fastapi",
  "typer",
  "uvicorn",
  # Database libraries
  "alembic",
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  # Project libraries
  "arxiv",
  "beautifulsoup4",
  "openai",
  "pypdf",
  "scipy",
  "streamlit",
  "tiktoken",
  "wikipedia",
  # Formatting
  "black",
  # Type checking
  "mypy==1.4.1",
  # Testing
  "pytest",
  # Linting
  "ruff",
  # phidata
  "phidata==2.0.0.dev28",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
exclude = ["llmenv*", ".venv*"]

[tool.mypy]
plugins = ["pydantic.mypy"]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
exclude = ["llmenv*", ".venv*"]

[[tool.mypy.overrides]]
module = ["phi.*", "pgvector.*", "setuptools.*"]
ignore_missing_imports = true
